7856 Move Woo UTM provider to its own file
