+++ WooFoundation/WooFoundation/Utilities/UTMParameters.swift
@@ -35,22 +35,3 @@ public extension UTMParametersProviding {
         return components.url
     }
 }
-
-public struct WooCommerceComUTMParametersProvider: UTMParametersProviding {
-    public let parameters: [UTMParameterKey: String?]
-
-    public init(campaign: String,
-         source: String,
-         content: String?) {
-        parameters = [
-            .medium: Constants.wooCommerceComUtmMedium,
-            .campaign: campaign,
-            .source: source,
-            .content: content
-        ]
-    }
-
-    private enum Constants {
-        static let wooCommerceComUtmMedium = "woo_ios"
-    }
-}
